requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
sqlalchemy==2.0.17
psycopg2-binary==2.9.6
pydantic==1.10.7
python-dotenv==0.21.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.10.2
pytest==7.2.2
httpx==0.23.0
bcrypt==4.0.1